Added initial files via upload
Added original R code given to me along with a version "analysisEXP.r" which contains some file path changes for my computer. Also uploaded the "Data.csv" which contains the output of our Matlab analysis along with the output from the "analysisEXP.r" done with that file. "COMBINED4.csv" is all the outputs combined into one file with the mappings and some values deleted or highlighted in the main spreadsheet.